feat(time): Months/Weekdays enums and Duration-typed duration constants
Rebuild the time module as StdModuleData{Consts, Funcs} with explicit buckets:
types, layouts, the new Months/Weekdays enums and the duration constants live in
the read-only Consts bucket, functions in Funcs. Month/weekday values are grouped
into `Months` / `Weekdays` enums (`time.Months.January`, `time.Weekdays.Sunday`)
instead of loose ints, and Nanosecond…Hour are now `duration` values rather than
ints.

To keep durations ergonomic: Int*Duration scales a duration (commutative, e.g.
`30 * time.Minute`) and ToGoInt64 accepts a Duration, so the int-typed duration
functions still take a duration argument. Time tests updated for the enum and
duration types.

+++ encoder/bytecode_test.go
@@ -75,7 +75,7 @@ v = int(strings.join([v], ""))
 v = srcmod.Incr(v)
 v = srcmod.Decr(v)
 v = int(fmt.sprintf("%d", v))
-return v*time.Second/time.Second // 1
+return int(v*time.Second/time.Second) // 1 (duration/duration is a float ratio)
 `
 
 	opts := gad.DefaultCompilerOptions
